+++ src/app/app.module.ts
@@ -71,6 +71,9 @@ import { TicketEditFormComponent } from './box-office/ticket-edit-form/ticket-ed
 import { SpotRegistrationComponent } from './box-office/registration/spot-registration/spot-registration.component';
 import { BulkUploadComponent } from './box-office/bookings/bulk-upload/bulk-upload.component';
 import { MyDepartmentComponent } from './my-department/my-department.component';
+import { EventsComponent } from './admin/events/events.component';
+import { EventDetailComponent } from './admin/events/event-detail/event-detail.component';
+import { EventCreateDialogComponent } from './admin/events/event-create-dialog/event-create-dialog.component';
 
 // #QRCode Scanner: Necessary to solve the problem of losing internet connection
 LOAD_WASM().subscribe()
@@ -115,7 +118,10 @@ LOAD_WASM().subscribe()
     TicketEditFormComponent,
     SpotRegistrationComponent,
     BulkUploadComponent,
-    MyDepartmentComponent
+    MyDepartmentComponent,
+    EventsComponent,
+    EventDetailComponent,
+    EventCreateDialogComponent
   ],
   imports: [
     BrowserModule,